requirements.txt
appdirs==1.4.3
asgi-redis==1.0.0
asgiref==1.0.0
autobahn==0.17.1
beautifulsoup4==4.5.1
channels==1.0.3
click==6.6
constantly==15.1.0
coreapi==2.3.1
coreapi-cli==1.0.6
coreschema==0.0.4
coverage==4.2
daphne==1.0.2
decorator==4.0.10
deps==0.1.0
dj-database-url==0.4.1
Django==1.10.4
django-autoslug==1.9.3
django-bootstrap-breadcrumbs==0.8
django-braces==1.10.0
django-common-helpers==0.9.0
django-cron==0.5.0
django-crontab==0.7.1
django-discover-runner==1.0
django-filter==1.0.4
django-floppyforms==1.7.0
django-modalview==0.1.5
django-oauth-toolkit==1.0.0
django-oauth2-provider==0.2.6.1
django-rest-swagger==2.1.2
django-role-permissions==1.2.1
django-s3direct==0.4.2
django-session-security==2.4.0
django-summernote==0.8.6
django-tracking2==0.3.3
django-unixdatetimefield==0.1.6
django-widget-tweaks==1.4.1
django-wysiwyg==0.8.0
djangoajax==2.3.6
djangorestframework==3.6.4
entrypoints==0.2.2
et-xmlfile==1.0.1
fcm-django==0.2.11
file-resubmit==0.1.0
gunicorn==19.6.0
httplib2==0.9.2
incremental==16.10.1
ipykernel==4.3.1
ipython==4.2.1
ipython-genutils==0.1.0
ipywidgets==5.1.5
itsdangerous==0.24
itypes==1.1.0
jdcal==1.3
Jinja2==2.8
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.3.0
jupyter-console==5.0.0
jupyter-core==4.1.0
lxml==3.7.3
Markdown==2.6.6
MarkupSafe==0.23
mistune==0.7.3
msgpack-python==0.4.8
nbconvert==4.2.0
nbformat==4.0.1
notebook==4.2.1
numpy==1.12.1
oauth2==1.9.0.post1
oauth2-provider==0.0
oauthlib==2.0.1
openapi-codec==1.3.2
openpyxl==2.4.5
packaging==16.8
pandas==0.19.2
pexpect==4.2.0
pickleshare==0.7.2
Pillow==3.3.1
prompt-toolkit==1.0.3
psycopg2==2.6.2
ptyprocess==0.5.1
pycpfcnpj==1.0.2
pyfcm==1.3.1
Pygments==2.1.3
pyparsing==2.2.0
python-dateutil==2.6.0
python-slugify==1.2.0
pytz==2016.10
pyzmq==15.3.0
qtconsole==4.2.1
redis==2.10.5
requests==2.13.0
requests-toolbelt==0.8.0
shortuuid==0.4.3
simplegeneric==0.8.1
simplejson==3.11.1
six==1.10.0
slugify==0.0.1
terminado==0.6
tornado==4.3
traitlets==4.2.2
Twisted==16.6.0
txaio==2.6.0
Unidecode==0.4.19
uritemplate==3.0.0
validators==0.11.0
virtualenv==15.0.3
wcwidth==0.1.7
webencodings==0.5
Werkzeug==0.11.11
whitenoise==3.2.2
widgetsnbextension==1.2.3
xlrd==1.0.0
xlwt==1.2.0
zope.interface==4.3.3
